[UT] Refactor some lake tests by using StarRocks mini cluster
