+++ fe/fe-core/src/test/java/com/starrocks/catalog/DropPartitionTest.java
@@ -34,39 +34,21 @@
 
 package com.starrocks.catalog;
 
-import com.google.common.collect.Lists;
-import com.staros.proto.AwsCredentialInfo;
-import com.staros.proto.AwsDefaultCredentialInfo;
-import com.staros.proto.FileCacheInfo;
-import com.staros.proto.FilePathInfo;
-import com.staros.proto.FileStoreInfo;
-import com.staros.proto.FileStoreType;
-import com.staros.proto.S3FileStoreInfo;
-import com.starrocks.common.AnalysisException;
 import com.starrocks.common.DdlException;
 import com.starrocks.common.ExceptionChecker;
-import com.starrocks.common.jmockit.Deencapsulation;
-import com.starrocks.lake.StarOSAgent;
 import com.starrocks.qe.ConnectContext;
 import com.starrocks.server.GlobalStateMgr;
-import com.starrocks.server.RunMode;
-import com.starrocks.server.SharedNothingStorageVolumeMgr;
 import com.starrocks.sql.ast.AlterTableStmt;
 import com.starrocks.sql.ast.CreateDbStmt;
 import com.starrocks.sql.ast.CreateTableStmt;
 import com.starrocks.sql.ast.RecoverPartitionStmt;
-import com.starrocks.storagevolume.StorageVolume;
 import com.starrocks.utframe.StarRocksAssert;
 import com.starrocks.utframe.UtFrameUtils;
-import mockit.Expectations;
-import mockit.Mock;
-import mockit.MockUp;
 import org.junit.Assert;
 import org.junit.BeforeClass;
 import org.junit.Test;
 
 import java.util.List;
-import java.util.Map;
 
 public class DropPartitionTest {
     private static ConnectContext connectContext;
@@ -86,74 +68,6 @@ public class DropPartitionTest {
                 + "buckets 1 properties('replication_num' = '1');";
         createDb(createDbStmtStr);
         createTable(createTableStr);
-
-        StarOSAgent agent = new StarOSAgent();
-
-        FilePathInfo.Builder builder = FilePathInfo.newBuilder();
-        FileStoreInfo.Builder fsBuilder = builder.getFsInfoBuilder();
-
-        S3FileStoreInfo.Builder s3FsBuilder = fsBuilder.getS3FsInfoBuilder();
-        s3FsBuilder.setBucket("test-bucket");
-        s3FsBuilder.setRegion("test-region");
-        s3FsBuilder.setCredential(AwsCredentialInfo.newBuilder()
-                .setDefaultCredential(AwsDefaultCredentialInfo.newBuilder().build()));
-        S3FileStoreInfo s3FsInfo = s3FsBuilder.build();
-
-        fsBuilder.setFsType(FileStoreType.S3);
-        fsBuilder.setFsKey("test-bucket");
-        fsBuilder.setFsName("test-fsname");
-        fsBuilder.setS3FsInfo(s3FsInfo);
-        FileStoreInfo fsInfo = fsBuilder.build();
-
-        builder.setFsInfo(fsInfo);
-        builder.setFullPath("s3://test-bucket/1/");
-        FilePathInfo pathInfo = builder.build();
-
-        new Expectations(agent) {
-            {
-                agent.allocateFilePath(anyString, anyLong, anyLong);
-                result = pathInfo;
-                agent.createShardGroup(anyLong, anyLong, anyLong);
-                result = GlobalStateMgr.getCurrentState().getNextId();
-                agent.createShards(anyInt, (FilePathInfo) any, (FileCacheInfo) any, anyLong, (Map<String, String>) any);
-                returns(Lists.newArrayList(10001L, 10002L, 10003L),
-                        Lists.newArrayList(10004L, 10005L, 10006L),
-                        Lists.newArrayList(10007L, 10008L, 10009L));
-                agent.getPrimaryComputeNodeIdByShard(anyLong, anyLong);
-                result = GlobalStateMgr.getCurrentSystemInfo().getBackendIds(true).get(0);
-            }
-        };
-
-        new MockUp<RunMode>() {
-            @Mock
-            public RunMode getCurrentRunMode() {
-                return RunMode.SHARED_DATA;
-            }
-        };
-
-        new MockUp<SharedNothingStorageVolumeMgr>() {
-            @Mock
-            public StorageVolume getStorageVolumeByName(String svName) throws AnalysisException {
-                return StorageVolume.fromFileStoreInfo(fsInfo);
-            }
-
-            @Mock
-            public String getStorageVolumeIdOfTable(long tableId) {
-                return fsInfo.getFsKey();
-            }
-        };
-
-        Deencapsulation.setField(GlobalStateMgr.getCurrentState(), "starOSAgent", agent);
-
-        String createLakeTableStr = "create table test.lake_table(k1 date, k2 int, k3 smallint, v1 varchar(2048), "
-                + "v2 datetime default '2014-02-04 15:36:00')"
-                + " duplicate key(k1, k2, k3)"
-                + " PARTITION BY RANGE(k1, k2, k3)"
-                + " (PARTITION p1 VALUES [(\"2014-01-01\", \"10\", \"200\"), (\"2014-01-01\", \"20\", \"300\")),"
-                + " PARTITION p2 VALUES [(\"2014-06-01\", \"100\", \"200\"), (\"2014-07-01\", \"100\", \"300\")))"
-                + " DISTRIBUTED BY HASH(k2) BUCKETS 3"
-                + " PROPERTIES ( \"datacache.enable\" = \"true\")";
-        createTable(createLakeTableStr);
     }
 
     private static void createDb(String sql) throws Exception {
@@ -226,47 +140,5 @@ public class DropPartitionTest {
         partition = table.getPartition("p20210203");
         Assert.assertEquals(1, replicaList.size());
         Assert.assertNull(partition);
-    }
-
-    @Test
-    public void testNormalDropLakePartition() throws Exception {
-        Database db = GlobalStateMgr.getCurrentState().getDb("test");
-        OlapTable table = (OlapTable) db.getTable("lake_table");
-        Partition partition = table.getPartition("p1");
-        long tabletId = partition.getBaseIndex().getTablets().get(0).getId();
-        String dropPartitionSql = " alter table test.lake_table drop partition p1;";
-        dropPartition(dropPartitionSql);
-        List<Replica> replicaList =
-                GlobalStateMgr.getCurrentState().getTabletInvertedIndex().getReplicasByTabletId(tabletId);
-        partition = table.getPartition("p1");
-        Assert.assertNull(partition);
-        String recoverPartitionSql = "recover partition p1 from test.lake_table";
-        RecoverPartitionStmt recoverPartitionStmt =
-                (RecoverPartitionStmt) UtFrameUtils.parseStmtWithNewParser(recoverPartitionSql, connectContext);
-        GlobalStateMgr.getCurrentState().recoverPartition(recoverPartitionStmt);
-        partition = table.getPartition("p1");
-        Assert.assertNotNull(partition);
-        Assert.assertEquals("p1", partition.getName());
-    }
-
-    @Test
-    public void testForceDropLakePartition() throws Exception {
-        Database db = GlobalStateMgr.getCurrentState().getDb("test");
-        OlapTable table = (OlapTable) db.getTable("lake_table");
-        Partition partition = table.getPartition("p1");
-        long tabletId = partition.getBaseIndex().getTablets().get(0).getId();
-        String dropPartitionSql = " alter table test.lake_table drop partition p1 force;";
-        dropPartition(dropPartitionSql);
-        List<Replica> replicaList =
-                GlobalStateMgr.getCurrentState().getTabletInvertedIndex().getReplicasByTabletId(tabletId);
-        partition = table.getPartition("p1");
-        Assert.assertTrue(replicaList.isEmpty());
-        Assert.assertNull(partition);
-        String recoverPartitionSql = "recover partition p1 from test.lake_table";
-        RecoverPartitionStmt recoverPartitionStmt =
-                (RecoverPartitionStmt) UtFrameUtils.parseStmtWithNewParser(recoverPartitionSql, connectContext);
-        ExceptionChecker.expectThrowsWithMsg(DdlException.class,
-                "No partition named p1 in table lake_table",
-                () -> GlobalStateMgr.getCurrentState().recoverPartition(recoverPartitionStmt));
     }
 }
