+++ fe/fe-core/src/test/java/com/starrocks/lake/LakeMaterializedViewTest.java
@@ -17,8 +17,6 @@ package com.starrocks.lake;
 import com.google.common.collect.Lists;
 import com.google.common.collect.Maps;
 import com.google.common.collect.Range;
-import com.staros.proto.AwsCredentialInfo;
-import com.staros.proto.AwsDefaultCredentialInfo;
 import com.staros.proto.FileCacheInfo;
 import com.staros.proto.FilePathInfo;
 import com.staros.proto.FileStoreInfo;
@@ -47,22 +45,17 @@ import com.starrocks.catalog.Tablet;
 import com.starrocks.catalog.TabletMeta;
 import com.starrocks.catalog.Type;
 import com.starrocks.common.AnalysisException;
-import com.starrocks.common.Config;
 import com.starrocks.common.FeConstants;
 import com.starrocks.common.io.FastByteArrayOutputStream;
 import com.starrocks.common.jmockit.Deencapsulation;
 import com.starrocks.common.util.PropertyAnalyzer;
 import com.starrocks.common.util.TimeUtils;
-import com.starrocks.pseudocluster.PseudoCluster;
 import com.starrocks.qe.ConnectContext;
 import com.starrocks.qe.StmtExecutor;
 import com.starrocks.scheduler.Task;
 import com.starrocks.server.GlobalStateMgr;
 import com.starrocks.server.RunMode;
-import com.starrocks.server.SharedDataStorageVolumeMgr;
-import com.starrocks.server.SharedNothingStorageVolumeMgr;
 import com.starrocks.sql.ast.AlterTableStmt;
-import com.starrocks.storagevolume.StorageVolume;
 import com.starrocks.thrift.TStorageMedium;
 import com.starrocks.thrift.TStorageType;
 import com.starrocks.utframe.StarRocksAssert;
@@ -87,14 +80,12 @@ import static com.starrocks.sql.optimizer.MVTestUtils.waitForSchemaChangeAlterJo
 public class LakeMaterializedViewTest {
     private static final String DB = "db_for_lake_mv";
 
-    private static PseudoCluster cluster;
     private static ConnectContext connectContext;
     private static StarRocksAssert starRocksAssert;
 
     @BeforeClass
     public static void setUp() throws Exception {
-        PseudoCluster.getOrCreateWithRandomPort(true, 3);
-        cluster = PseudoCluster.getInstance();
+        UtFrameUtils.createMinStarRocksCluster(RunMode.SHARED_DATA);
         connectContext = UtFrameUtils.createDefaultCtx();
 
         // set default config for async mvs
@@ -102,69 +93,6 @@ public class LakeMaterializedViewTest {
 
         starRocksAssert = new StarRocksAssert(connectContext);
         starRocksAssert.withDatabase(DB).useDatabase(DB);
-
-        new MockUp<StarOSAgent>() {
-            @Mock
-            public long getPrimaryComputeNodeIdByShard(long shardId, long workerGroupId) {
-                return GlobalStateMgr.getCurrentSystemInfo().getBackendIds(true).get(0);
-            }
-
-            @Mock
-            public FilePathInfo allocateFilePath(String storageVolumeId, long dbId, long tableId) {
-                FilePathInfo.Builder builder = FilePathInfo.newBuilder();
-                FileStoreInfo.Builder fsBuilder = builder.getFsInfoBuilder();
-
-                S3FileStoreInfo.Builder s3FsBuilder = fsBuilder.getS3FsInfoBuilder();
-                s3FsBuilder.setBucket("test-bucket");
-                s3FsBuilder.setRegion("test-region");
-                s3FsBuilder.setCredential(AwsCredentialInfo.newBuilder()
-                        .setDefaultCredential(AwsDefaultCredentialInfo.newBuilder().build()));
-                S3FileStoreInfo s3FsInfo = s3FsBuilder.build();
-
-                fsBuilder.setFsType(FileStoreType.S3);
-                fsBuilder.setFsKey("test-bucket");
-                fsBuilder.setFsName("test-fsname");
-                fsBuilder.setS3FsInfo(s3FsInfo);
-                FileStoreInfo fsInfo = fsBuilder.build();
-
-                builder.setFsInfo(fsInfo);
-                builder.setFullPath("s3://test-bucket/1/");
-                FilePathInfo pathInfo = builder.build();
-                return pathInfo;
-            }
-        };
-
-        new MockUp<RunMode>() {
-            @Mock
-            public RunMode getCurrentRunMode() {
-                return RunMode.SHARED_DATA;
-            }
-        };
-
-        new MockUp<SharedNothingStorageVolumeMgr>() {
-            S3FileStoreInfo s3FileStoreInfo = S3FileStoreInfo.newBuilder().setBucket("default-bucket")
-                    .setRegion(Config.aws_s3_region).setEndpoint(Config.aws_s3_endpoint)
-                    .setCredential(AwsCredentialInfo.newBuilder()
-                            .setDefaultCredential(AwsDefaultCredentialInfo.newBuilder().build()).build()).build();
-            FileStoreInfo fsInfo = FileStoreInfo.newBuilder().setFsName(SharedDataStorageVolumeMgr.BUILTIN_STORAGE_VOLUME)
-                    .setFsKey("1").setFsType(FileStoreType.S3)
-                    .setS3FsInfo(s3FileStoreInfo).build();
-
-            @Mock
-            public StorageVolume getStorageVolumeByName(String svName) throws AnalysisException {
-                return StorageVolume.fromFileStoreInfo(fsInfo);
-            }
-
-            @Mock
-            public StorageVolume getStorageVolume(String svKey) throws AnalysisException {
-                return StorageVolume.fromFileStoreInfo(fsInfo);
-            }
-
-            @Mock
-            public String getStorageVolumeIdOfTable(long tableId) {
-                return fsInfo.getFsKey();
-            }
-        };
 
         starRocksAssert.withTable("CREATE TABLE base_table\n" +
                 "(\n" +
@@ -182,7 +110,7 @@ public class LakeMaterializedViewTest {
 
     @AfterClass
     public static void tearDown() {
-        PseudoCluster.getInstance().shutdown(true);
+
     }
 
     @Test
@@ -326,7 +254,7 @@ public class LakeMaterializedViewTest {
 
         LakeMaterializedView lakeMv = (LakeMaterializedView) mv;
         // same as PseudoStarOSAgent.allocateFilePath
-        Assert.assertEquals("s3://test-bucket/1/", lakeMv.getDefaultFilePathInfo().getFullPath());
+        Assert.assertTrue(lakeMv.getDefaultFilePathInfo().getFullPath().startsWith("s3://dummy_unittest_bucket/dummy_sub_path"));
         // check table default cache info
         FileCacheInfo cacheInfo = lakeMv.getPartitionFileCacheInfo(0L);
         Assert.assertTrue(cacheInfo.getEnableCache());
